Better test tear down

+++ src/test/java/org/apache/commons/compress/AbstractTestCase.java
@@ -86,7 +86,7 @@ public abstract class AbstractTestCase {
      */
     public static boolean tryHardToDelete(final File file) {
         try {
-            if (file != null) {
+            if (file != null && file.exists()) {
                 FileUtils.forceDelete(file);
             }
             return true;
